Added test to ensure that rando * 0 yields expected results.
