+++ vanilla-core/src/test/groovy/com/stehno/vanilla/test/Person.groovy
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2016 Christopher J. Stehno <chris@stehno.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.stehno.vanilla.test
+
+import groovy.transform.EqualsAndHashCode
+import groovy.transform.ToString
+
+/**
+ * Created by cstehno on 1/11/2016.
+ */
+@ToString(includeNames = true) @EqualsAndHashCode
+class Person {
+    String name
+    int age
+    Date birthDate
+    Pet pet
+    String bankPin
+
+    String[] children
+
+    int getChildCount() {
+        children?.size() ?: 0
+    }
+}
